requirements.txt
flask
pillow
pytesseract
anthropic